+++ index.js
@@ -1,6 +1,11 @@
 const puppeteer = require('puppeteer');
 
 const {
+  basename,
+  resolve
+} = require('path');
+
+const {
   readFileSync
 } = require('fs');
 
@@ -10,16 +15,24 @@ async function printDiagram(page, options) {
   const {
     input,
     outputs,
-    minDimensions
+    minDimensions,
+    title
   } = options;
 
   const diagramXML = readFileSync(input, 'utf8');
 
+  const diagramTitle = title === false ? false : (
+    title.length ? title : basename(input)
+  );
+
   await page.goto(`file://${__dirname}/skeleton.html`);
 
-  const desiredViewport = await page.evaluate((diagramXML, minDimensions) => {
-    return openDiagram(diagramXML, minDimensions);
-  }, diagramXML, minDimensions);
+  const desiredViewport = await page.evaluate((diagramXML, minDimensions, title) => {
+    return openDiagram(diagramXML, {
+      minDimensions,
+      title
+    });
+  }, diagramXML, minDimensions, diagramTitle);
 
   page.setViewport({
     width: desiredViewport.width,
@@ -83,7 +96,8 @@ async function withPage(fn) {
 module.exports.convertAll = async function(conversions, options={}) {
 
   const {
-    minDimensions
+    minDimensions,
+    title
   } = options;
 
   await withPage(async function(page) {
@@ -98,7 +112,8 @@ module.exports.convertAll = async function(conversions, options={}) {
       await printDiagram(page, {
         input,
         outputs,
-        minDimensions
+        minDimensions,
+        title
       });
     }